Moving the lab card examples to a new file
