+++ utils/lab-uikit.ts
@@ -72,7 +72,3 @@ export const canLabWindowGroup = (scene: Scene) => {
 
   return windowGroupMesh;
 };
-
-// export const canLabToolbar = (scene: Scene) => {
-
-// };
